Avoid warning in tests that the path being removed doesn't exist.
This function is used to ensure a clean state, not to check that
the previous state is as expected, so erroring about a missing path
is just a distraction here.

+++ tests/test-header.php
@@ -164,6 +164,12 @@ function subsection($label)
 
 function cleanCreate($path)
 {
-	delete_recursive($path);
+	// while it's not nice to have conditionals in tests,
+	// this avoids an error about the target not existing
+	// without hiding other failure modes of the deletion.
+	if (file_exists($path))
+	{
+		delete_recursive($path);
+	}
 	mkdir_full($path);
 }
